hexdump, vfs and initrd
Made two tools: hexdump and mk-initrd
The hexdump does the same as the Unix hexdump does (with the -C flag).
mk-initrd is for creating the initial ramdisk image, it takes pathnames
as arguments.

The virtual file system implementation is at some stage, reading files
from the initrd works. The initrd is loaded by grub as a module, hence the
multiboot.h.