Add sys/stat.h, unify rlibc and kernel headers, typedef system better
Kernel and rlibc now share a bunch of headers (in razos_kernel_headers).
Reason for this is trying to avoid code duplication and make maintaining
easier. The shared headers include platform-specific info, such as types,
and all kinds of flags that must have same values in the C library and
in the kernel. This lead to quite complex macro system, be careful!

Add sys/stat.h. On the way to a better virtua file system, starting with
file status information. FIFOs and pipes are on the way.