shell works (enough), shutdown somewhat works, fix strcmp
Shutdown routine is quite crappy (we just hope acpi is already initialized,
as in bochs and qemu), but it works..

There seems to be some problems with memory handling in the master's head,
so this branch is based on the old (black magic based) memory manager.

strcmp, strncmp were faulty, they are fixed now.